+++ src/_lcm/solution/solvers.py
@@ -1,0 +1,250 @@
+"""The built-in regime solvers.
+
+`GridSearch` (the default) runs the existing max-Q-over-a grid search;
+`DCEGM` is a published configuration whose engine is not yet wired in, so a
+regime requesting it is rejected at model build by its `validate`. Both are
+`Solver` subclasses. The kernel-building imports (`jax`, `get_max_Q_over_a`)
+are function-local so the public `lcm.solvers` façade stays a thin re-export
+that pulls in no numerical engine modules.
+"""
+
+import math
+from dataclasses import dataclass
+from types import MappingProxyType
+from typing import Literal
+
+from beartype import beartype
+
+from _lcm.beartype_conf import REGIME_CONF
+from _lcm.grids import ContinuousGrid
+from _lcm.processes.base import _ContinuousStochasticProcess
+from _lcm.solution.contract import Solver, SolverBuildContext, SolverKernels
+from _lcm.typing import MaxQOverAFunction
+from lcm.exceptions import RegimeInitializationError
+from lcm.typing import ActionName, FunctionName, StateName
+
+
+@beartype(conf=REGIME_CONF)
+@dataclass(frozen=True, kw_only=True)
+class GridSearch(Solver):
+    """Grid-search solver over the full state-action product (the default)."""
+
+    def build_period_kernels(self, *, context: SolverBuildContext) -> SolverKernels:
+        """Build max-Q-over-a closures for each period.
+
+        Periods sharing the same Q_and_F object reuse a single compiled
+        function.
+        """
+        import jax  # noqa: PLC0415
+
+        from _lcm.regime_building.max_Q_over_a import get_max_Q_over_a  # noqa: PLC0415
+
+        built: dict[int, MaxQOverAFunction] = {}
+        result: dict[int, MaxQOverAFunction] = {}
+        for period, Q_and_F in context.Q_and_F_functions.items():
+            q_id = id(Q_and_F)
+            if q_id not in built:
+                func = get_max_Q_over_a(
+                    Q_and_F=Q_and_F,
+                    batch_sizes={
+                        name: grid.batch_size
+                        for name, grid in context.grids.items()
+                        if name in context.state_action_space.state_names
+                    },
+                    action_names=context.state_action_space.action_names,
+                    state_names=context.state_action_space.state_names,
+                    n_discrete_action_axes=len(
+                        context.state_action_space.discrete_actions
+                    ),
+                    has_taste_shocks=context.has_taste_shocks,
+                )
+                built[q_id] = jax.jit(func) if context.enable_jit else func
+            result[period] = built[q_id]
+        return SolverKernels(max_Q_over_a=MappingProxyType(result))
+
+
+@beartype(conf=REGIME_CONF)
+@dataclass(frozen=True, kw_only=True)
+class DCEGM(Solver):
+    """Configuration of the DC-EGM solver for one regime.
+
+    DC-EGM inverts the Euler equation on an exogenous end-of-period
+    (post-decision) grid instead of searching a dense grid for the continuous
+    action. It requires a specific model structure — exactly one continuous
+    (*Euler*) state and one continuous action, a declared resources function
+    `R` with consumption recovery `c = R - A`, a post-decision function `A`,
+    and an `inverse_marginal_utility` regime function — which is validated at
+    `Model` construction time.
+
+    The configuration is published so a model can name the solver and its
+    parameters, but the solver engine is not yet wired in: `validate` rejects a
+    regime requesting it. `GridSearch` is the only available solver.
+
+    Forward simulation works but is *grid-restricted*: `simulate` recomputes
+    the argmax over the regime's gridded continuous action against the
+    stored value function, rather than interpolating the exact EGM policy.
+    Simulated continuous actions therefore live on the action grid, and with
+    taste shocks the simulated choice frequencies follow the grid-restricted
+    choice-specific values, not exactly the solve's choice probabilities.
+    The budget constraint the solve enforces intrinsically
+    (`continuous_action <= resources - savings_grid lower bound`) is applied
+    as a feasibility mask during simulation.
+
+    """
+
+    continuous_state: StateName
+    """Name of the Euler continuous state (e.g. `"wealth"`).
+
+    Its transition must consume the post-decision function and reach the
+    state and the continuous action only through it.
+    """
+
+    continuous_action: ActionName
+    """Name of the continuous action (e.g. `"consumption"`)."""
+
+    resources: FunctionName
+    """Name of the resources function `R` in `Regime.functions`.
+
+    Resources are what consumption is paid out of; the endogenous grid lives
+    in R-space. Required even in the classic case, where it is the identity
+    (e.g. `"resources": lambda wealth: wealth`). Must not depend on the
+    continuous action and must be non-decreasing in the continuous state.
+    """
+
+    post_decision_function: FunctionName
+    """Name of the post-decision function in `Regime.functions`.
+
+    The end-of-period state (e.g. savings), satisfying
+    `post_decision = resources - continuous_action`.
+    """
+
+    savings_grid: ContinuousGrid
+    """Exogenous end-of-period grid; its lower bound is the borrowing limit.
+
+    The endogenous grid inherits this grid's spacing, and the published value
+    function is interpolated linearly between endogenous points — so this grid
+    controls where the solution is accurate. With sharply curved utility (e.g.
+    CRRA), cluster the nodes toward the borrowing limit (`LogSpacedGrid`, or
+    an `IrregSpacedGrid` clustered at the low end): a uniform grid
+    under-resolves the value function near the limit, and that interpolation
+    error compounds across periods.
+    """
+
+    upper_envelope: Literal["fues"] = "fues"
+    """Upper-envelope refinement backend removing dominated Euler candidates."""
+
+    fues_jump_thresh: float = 2.0
+    """Segment-switch threshold on `|ΔA / ΔR|` in the FUES scan."""
+
+    fues_n_points_to_scan: int = 10
+    """Number of points the FUES forward scan inspects after a candidate."""
+
+    refined_grid_factor: float = 1.2
+    """Headroom factor sizing the refined (NaN-padded) envelope arrays."""
+
+    n_constrained_points: int = 20
+    """Number of closed-form points on the credit-constrained segment."""
+
+    stochastic_node_batch_size: int = 0
+    """Block size for splaying the child stochastic-node expectation.
+
+    The continuation expectation runs over the product of the child regime's
+    stochastic process nodes — a single mesh, not a per-grid axis, so it gets
+    its own solve-level knob rather than a per-grid `batch_size`. A positive
+    value below the mesh length processes that expectation in `lax.map` blocks
+    instead of one fused vmap, shedding the dominant `egm_step` working buffer
+    (which carries this node axis); `0` keeps the fused vmap. Like the savings
+    grid's `batch_size`, this is a memory knob only — the solved value function
+    is identical to the unsplayed solve.
+    """
+
+    def __post_init__(self) -> None:
+        _fail_if_savings_grid_is_stochastic(self.savings_grid)
+        _fail_if_refined_grid_factor_too_small(self.refined_grid_factor)
+        _fail_if_fues_jump_thresh_non_positive(self.fues_jump_thresh)
+        _fail_if_n_constrained_points_too_few(self.n_constrained_points)
+        _fail_if_fues_n_points_to_scan_too_few(self.fues_n_points_to_scan)
+        _fail_if_stochastic_node_batch_size_negative(self.stochastic_node_batch_size)
+
+    def validate(self, *, context: SolverBuildContext) -> None:
+        """Reject the not-yet-available DC-EGM solver at model build."""
+        msg = (
+            "The DC-EGM solver is not yet available. A regime requests "
+            "`solver=DCEGM(...)`; use `GridSearch()` (the default) until the "
+            "DC-EGM engine is wired in."
+        )
+        raise NotImplementedError(msg)
+
+    def build_period_kernels(self, *, context: SolverBuildContext) -> SolverKernels:
+        """Unreachable — `validate` rejects DC-EGM before the kernel build."""
+        msg = "The DC-EGM solver is not yet available."
+        raise NotImplementedError(msg)
+
+
+def _fail_if_savings_grid_is_stochastic(savings_grid: ContinuousGrid) -> None:
+    if isinstance(savings_grid, _ContinuousStochasticProcess):
+        msg = (
+            "DCEGM.savings_grid must be a deterministic continuous grid, not a "
+            f"stochastic process ({type(savings_grid).__name__}). The savings "
+            "grid is the exogenous end-of-period grid; it carries no transition."
+        )
+        raise RegimeInitializationError(msg)
+
+
+def _fail_if_refined_grid_factor_too_small(refined_grid_factor: float) -> None:
+    # `not (x > 1.0)` rejects NaN too — `nan <= 1.0` is False, so a bare
+    # `<= 1.0` guard would admit a non-finite factor that later sizes the
+    # refined envelope arrays and corrupts the scatter.
+    if not (math.isfinite(refined_grid_factor) and refined_grid_factor > 1.0):
+        msg = (
+            f"DCEGM.refined_grid_factor must be a finite value greater than 1.0, "
+            f"got {refined_grid_factor}. It is the headroom factor sizing the "
+            "refined envelope arrays; a value at or below 1.0 leaves no room "
+            "for the constrained points and overflows the scatter."
+        )
+        raise RegimeInitializationError(msg)
+
+
+def _fail_if_fues_jump_thresh_non_positive(fues_jump_thresh: float) -> None:
+    # `not (x > 0.0)` rejects NaN too: `nan <= 0.0` is False, so the segment-
+    # switch comparison would silently misbehave on a non-finite threshold.
+    if not (math.isfinite(fues_jump_thresh) and fues_jump_thresh > 0.0):
+        msg = (
+            f"DCEGM.fues_jump_thresh must be a finite positive value, got "
+            f"{fues_jump_thresh}. It is the segment-switch threshold on "
+            "`|ΔA / ΔR|` in the FUES scan."
+        )
+        raise RegimeInitializationError(msg)
+
+
+def _fail_if_n_constrained_points_too_few(n_constrained_points: int) -> None:
+    if n_constrained_points < 2:  # noqa: PLR2004
+        msg = (
+            f"DCEGM.n_constrained_points must be at least 2, got "
+            f"{n_constrained_points}. The credit-constrained segment needs at "
+            "least two closed-form points to interpolate between."
+        )
+        raise RegimeInitializationError(msg)
+
+
+def _fail_if_fues_n_points_to_scan_too_few(fues_n_points_to_scan: int) -> None:
+    if fues_n_points_to_scan < 1:
+        msg = (
+            f"DCEGM.fues_n_points_to_scan must be at least 1, got "
+            f"{fues_n_points_to_scan}. The FUES forward scan must inspect at "
+            "least one point after each candidate."
+        )
+        raise RegimeInitializationError(msg)
+
+
+def _fail_if_stochastic_node_batch_size_negative(
+    stochastic_node_batch_size: int,
+) -> None:
+    if stochastic_node_batch_size < 0:
+        msg = (
+            f"DCEGM.stochastic_node_batch_size must be non-negative, got "
+            f"{stochastic_node_batch_size}. It is the block size for splaying the "
+            "child stochastic-node expectation into `lax.map` blocks; 0 keeps the "
+            "fused vmap."
+        )
+        raise RegimeInitializationError(msg)
