+++ src/lcm/solvers.py
@@ -1,202 +1,28 @@
-"""User-facing solver configuration classes.
+"""User-facing solver configuration (re-export façade).
 
 A regime's `solver` field selects the algorithm used for that regime during
 backward induction:
 
-- `BruteForce()` (the default): grid search over the full state-action product.
+- `GridSearch()` (the default): grid search over the full state-action product.
 - `DCEGM(...)`: the endogenous grid method for discrete-continuous choice
   (Iskhakov, Jørgensen, Rust & Schjerning 2017, Quantitative Economics 8(2),
-  317-365, [doi:10.3982/QE643](https://doi.org/10.3982/QE643)).
+  317-365, [doi:10.3982/QE643](https://doi.org/10.3982/QE643)) — published, but
+  its engine is not yet wired in.
 
-These classes are pure configuration — no callables, no engine internals. The
-engine dispatches on `type(regime.solver)` through the private builder
-registry in `_lcm.solution.registry`.
-
+The solvers are defined engine-side in `_lcm.solution.solvers`; this module is a
+thin re-export so user code (and `lcm.regime`) can name them, and the `Solver`
+contract, without eagerly importing the numerical engine. The engine dispatches
+polymorphically on the solver instance (`solver.build_period_kernels(context)`),
+not on its type.
 """
 
-import math
-from dataclasses import dataclass
-from typing import Literal
+from _lcm.solution.contract import Solver, SolverBuildContext, SolverKernels
+from _lcm.solution.solvers import DCEGM, GridSearch
 
-from beartype import beartype
-
-from _lcm.beartype_conf import REGIME_CONF
-from _lcm.grids import ContinuousGrid
-from _lcm.processes.base import _ContinuousStochasticProcess
-from lcm.exceptions import RegimeInitializationError
-from lcm.typing import ActionName, FunctionName, StateName
-
-
-@beartype(conf=REGIME_CONF)
-@dataclass(frozen=True, kw_only=True)
-class BruteForce:
-    """Grid-search solver over the full state-action product (the default)."""
-
-
-@beartype(conf=REGIME_CONF)
-@dataclass(frozen=True, kw_only=True)
-class DCEGM:
-    """Configuration of the DC-EGM solver for one regime.
-
-    DC-EGM inverts the Euler equation on an exogenous end-of-period
-    (post-decision) grid instead of searching a dense grid for the continuous
-    action. It requires a specific model structure — exactly one continuous
-    (*Euler*) state and one continuous action, a declared resources function
-    `R` with consumption recovery `c = R - A`, a post-decision function `A`,
-    and an `inverse_marginal_utility` regime function — which is validated at
-    `Model` construction time.
-
-    Forward simulation works but is *grid-restricted*: `simulate` recomputes
-    the argmax over the regime's gridded continuous action against the
-    stored value function, rather than interpolating the exact EGM policy.
-    Simulated continuous actions therefore live on the action grid, and with
-    taste shocks the simulated choice frequencies follow the grid-restricted
-    choice-specific values, not exactly the solve's choice probabilities.
-    The budget constraint the solve enforces intrinsically
-    (`continuous_action <= resources - savings_grid lower bound`) is applied
-    as a feasibility mask during simulation.
-
-    """
-
-    continuous_state: StateName
-    """Name of the Euler continuous state (e.g. `"wealth"`).
-
-    Its transition must consume the post-decision function and reach the
-    state and the continuous action only through it.
-    """
-
-    continuous_action: ActionName
-    """Name of the continuous action (e.g. `"consumption"`)."""
-
-    resources: FunctionName
-    """Name of the resources function `R` in `Regime.functions`.
-
-    Resources are what consumption is paid out of; the endogenous grid lives
-    in R-space. Required even in the classic case, where it is the identity
-    (e.g. `"resources": lambda wealth: wealth`). Must not depend on the
-    continuous action and must be non-decreasing in the continuous state.
-    """
-
-    post_decision_function: FunctionName
-    """Name of the post-decision function in `Regime.functions`.
-
-    The end-of-period state (e.g. savings), satisfying
-    `post_decision = resources - continuous_action`.
-    """
-
-    savings_grid: ContinuousGrid
-    """Exogenous end-of-period grid; its lower bound is the borrowing limit.
-
-    The endogenous grid inherits this grid's spacing, and the published value
-    function is interpolated linearly between endogenous points — so this grid
-    controls where the solution is accurate. With sharply curved utility (e.g.
-    CRRA), cluster the nodes toward the borrowing limit (`LogSpacedGrid`, or
-    an `IrregSpacedGrid` clustered at the low end): a uniform grid
-    under-resolves the value function near the limit, and that interpolation
-    error compounds across periods.
-    """
-
-    upper_envelope: Literal["fues"] = "fues"
-    """Upper-envelope refinement backend removing dominated Euler candidates."""
-
-    fues_jump_thresh: float = 2.0
-    """Segment-switch threshold on `|ΔA / ΔR|` in the FUES scan."""
-
-    fues_n_points_to_scan: int = 10
-    """Number of points the FUES forward scan inspects after a candidate."""
-
-    refined_grid_factor: float = 1.2
-    """Headroom factor sizing the refined (NaN-padded) envelope arrays."""
-
-    n_constrained_points: int = 20
-    """Number of closed-form points on the credit-constrained segment."""
-
-    stochastic_node_batch_size: int = 0
-    """Block size for splaying the child stochastic-node expectation.
-
-    The continuation expectation runs over the product of the child regime's
-    stochastic process nodes — a single mesh, not a per-grid axis, so it gets
-    its own solve-level knob rather than a per-grid `batch_size`. A positive
-    value below the mesh length processes that expectation in `lax.map` blocks
-    instead of one fused vmap, shedding the dominant `egm_step` working buffer
-    (which carries this node axis); `0` keeps the fused vmap. Like the savings
-    grid's `batch_size`, this is a memory knob only — the solved value function
-    is identical to the unsplayed solve.
-    """
-
-    def __post_init__(self) -> None:
-        _fail_if_savings_grid_is_stochastic(self.savings_grid)
-        _fail_if_refined_grid_factor_too_small(self.refined_grid_factor)
-        _fail_if_fues_jump_thresh_non_positive(self.fues_jump_thresh)
-        _fail_if_n_constrained_points_too_few(self.n_constrained_points)
-        _fail_if_fues_n_points_to_scan_too_few(self.fues_n_points_to_scan)
-        _fail_if_stochastic_node_batch_size_negative(self.stochastic_node_batch_size)
-
-
-def _fail_if_savings_grid_is_stochastic(savings_grid: ContinuousGrid) -> None:
-    if isinstance(savings_grid, _ContinuousStochasticProcess):
-        msg = (
-            "DCEGM.savings_grid must be a deterministic continuous grid, not a "
-            f"stochastic process ({type(savings_grid).__name__}). The savings "
-            "grid is the exogenous end-of-period grid; it carries no transition."
-        )
-        raise RegimeInitializationError(msg)
-
-
-def _fail_if_refined_grid_factor_too_small(refined_grid_factor: float) -> None:
-    # `not (x > 1.0)` rejects NaN too — `nan <= 1.0` is False, so a bare
-    # `<= 1.0` guard would admit a non-finite factor that later sizes the
-    # refined envelope arrays and corrupts the scatter.
-    if not (math.isfinite(refined_grid_factor) and refined_grid_factor > 1.0):
-        msg = (
-            f"DCEGM.refined_grid_factor must be a finite value greater than 1.0, "
-            f"got {refined_grid_factor}. It is the headroom factor sizing the "
-            "refined envelope arrays; a value at or below 1.0 leaves no room "
-            "for the constrained points and overflows the scatter."
-        )
-        raise RegimeInitializationError(msg)
-
-
-def _fail_if_fues_jump_thresh_non_positive(fues_jump_thresh: float) -> None:
-    # `not (x > 0.0)` rejects NaN too: `nan <= 0.0` is False, so the segment-
-    # switch comparison would silently misbehave on a non-finite threshold.
-    if not (math.isfinite(fues_jump_thresh) and fues_jump_thresh > 0.0):
-        msg = (
-            f"DCEGM.fues_jump_thresh must be a finite positive value, got "
-            f"{fues_jump_thresh}. It is the segment-switch threshold on "
-            "`|ΔA / ΔR|` in the FUES scan."
-        )
-        raise RegimeInitializationError(msg)
-
-
-def _fail_if_n_constrained_points_too_few(n_constrained_points: int) -> None:
-    if n_constrained_points < 2:  # noqa: PLR2004
-        msg = (
-            f"DCEGM.n_constrained_points must be at least 2, got "
-            f"{n_constrained_points}. The credit-constrained segment needs at "
-            "least two closed-form points to interpolate between."
-        )
-        raise RegimeInitializationError(msg)
-
-
-def _fail_if_fues_n_points_to_scan_too_few(fues_n_points_to_scan: int) -> None:
-    if fues_n_points_to_scan < 1:
-        msg = (
-            f"DCEGM.fues_n_points_to_scan must be at least 1, got "
-            f"{fues_n_points_to_scan}. The FUES forward scan must inspect at "
-            "least one point after each candidate."
-        )
-        raise RegimeInitializationError(msg)
-
-
-def _fail_if_stochastic_node_batch_size_negative(
-    stochastic_node_batch_size: int,
-) -> None:
-    if stochastic_node_batch_size < 0:
-        msg = (
-            f"DCEGM.stochastic_node_batch_size must be non-negative, got "
-            f"{stochastic_node_batch_size}. It is the block size for splaying the "
-            "child stochastic-node expectation into `lax.map` blocks; 0 keeps the "
-            "fused vmap."
-        )
-        raise RegimeInitializationError(msg)
+__all__ = [
+    "DCEGM",
+    "GridSearch",
+    "Solver",
+    "SolverBuildContext",
+    "SolverKernels",
+]
